extract contentfetcher implementation
