+++ composer/src/main/java/com/rewedigital/examples/msintegration/composer/proxy/ComposerFactory.java
@@ -4,12 +4,13 @@ import java.util.Map;
 
 import com.rewedigital.examples.msintegration.composer.composing.AttoParserBasedComposer;
 import com.rewedigital.examples.msintegration.composer.composing.TemplateComposer;
+import com.rewedigital.examples.msintegration.composer.composing.ValidatingContentFetcher;
 import com.spotify.apollo.Client;
 
 public class ComposerFactory {
 
     public TemplateComposer build(final Client client, final Map<String, Object> parsedPathArguments) {
-        return new AttoParserBasedComposer(client, parsedPathArguments);
+        return new AttoParserBasedComposer(new ValidatingContentFetcher(client, parsedPathArguments));
     }
 
 }
